renamed exponentialmoving back to ewm.py
